Rename /space to /space_info
