+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - `/public_upload_photo`, `/public_upload_file`, `/public_upload_audio`, `/public_upload_video`, `/public_upload_voice`, `/public_upload_url`: uploading of files and then publishing them.
 - `/publish`: publishing of files or folders.
 - `/unpublish`: unpublishing of files or folders.
-- `/space`: getting of information about remaining Yandex.Disk space.
+- `/space_info`: getting of information about remaining Yandex.Disk space.
 - `/element_info`: getting of information about file or folder.
 - `/disk_info`: getting of information about Yandex.Disk.
 - Handling of files that exceed file size limit in 20 MB. At the moment the bot will asnwer with warning message, not trying to upload that file. [#3](https://github.com/Amaimersion/yandex-disk-telegram-bot/issues/3)
